+++ partiql-value/src/bag.rs
@@ -8,7 +8,7 @@ use std::hash::{Hash, Hasher};
 
 use std::{slice, vec};
 
-use crate::{List, NullSortedValue, Value};
+use crate::{EqualityValue, List, NullSortedValue, NullableEq, Value};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
@@ -176,10 +176,13 @@ impl PartialEq for Bag {
         if self.len() != other.len() {
             return false;
         }
-
-        let lhs = self.0.iter().sorted();
-        let rhs = other.0.iter().sorted();
-        lhs.zip(rhs).all(|(l, r)| l == r)
+        for (v1, v2) in self.0.iter().sorted().zip(other.0.iter().sorted()) {
+            let wrap = EqualityValue::<true, Value>;
+            if NullableEq::eq(&wrap(v1), &wrap(v2)) != Value::Boolean(true) {
+                return false;
+            }
+        }
+        true
     }
 }